Implement all and any methods
