--- MyEnumerable.rb
@@ -1,13 +1,0 @@
-module MyEnumerable
-  def all?(e)
-    # code
-  end
-  
-  def any?(e)
-    # code
-  end
-
-  def filter(e)
-    # code
-  end
-end
