+++ my_enumerable.rb
@@ -1,0 +1,19 @@
+module MyEnumerable
+  def all?
+    each do |pattern|
+      return false unless yield(pattern)
+    end
+    true
+  end
+
+  def any?
+    each do |pattern|
+      return true if yield(pattern)
+    end
+    false
+  end
+
+  def filter?
+    # code
+  end
+end
