+++ my_list.rb
@@ -1,0 +1,20 @@
+require_relative 'my_enumerable'
+
+class MyList
+  include MyEnumerable
+
+  def initialize(*elements)
+    @list = elements
+  end
+
+  def each(&block)
+    @list.each(&block)
+  end
+end
+
+list = MyList.new(1, 2, 3, 4)
+puts list.all? { |e| e < 5 }
+puts list.all? { |e| e > 5 }
+
+puts list.any? { |e| e == 2 }
+puts list.any? { |e| e == 5 }
